+++ thoth/python/__init__.py
@@ -27,7 +27,7 @@ from .pipfile import PipfileLock
 from .project import Project
 from .source import Source
 from .aiosource import AIOSource, AsyncIterablePackages, AsyncIterableVersions, AsyncIterableArtifacts
-from .lazy import Lazy
+from .helpers import Lazy
 
 
 __version__ = "0.9.2"
