+++ thoth/python/helpers.py
@@ -186,3 +186,14 @@ def parse_requirements(file_path: str) -> Tuple[List[Source], List[PackageVersio
         )
 
     return sources, package_versions
+
+class Lazy(object):
+    def __init__(self, calculate_function):
+        self._calculate = calculate_function
+
+    def __get__(self, obj, _=None):
+        if obj is None:
+            return self
+        value = self._calculate(obj)
+        setattr(obj, self._calculate.__name__, value)
+        return value
