--- thoth/python/lazy.py
@@ -1,10 +1,0 @@
-class Lazy(object):
-    def __init__(self, calculate_function):
-        self._calculate = calculate_function
-
-    def __get__(self, obj, _=None):
-        if obj is None:
-            return self
-        value = self._calculate(obj)
-        setattr(obj, self._calculate.__name__, value)
-        return value
